# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Unnamed URCM",
        "title": "Unnamed URCM",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/dashboards": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Get user`s dashboards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Dashboard"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/addAccess": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Add Access",
                "parameters": [
                    {
                        "description": " ",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddAccessPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/custom-field/{fieldId}": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Delete custom field",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "fieldId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/removeAccess/{dashboardId}/{userId}": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Remove access",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "dashboardId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/updateAccess": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Update access",
                "parameters": [
                    {
                        "description": " ",
                        "name": "AddAccessPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/{dashboardId}": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Delete dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "dashboardId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Update dashboard name",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "dashboardId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateNamePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/{dashboardId}/settings": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Add webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "dashboardId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddSettingsPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DashboardSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/admin/{dashboardId}/webhook": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Add webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "dashboardId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateNamePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
        "/dashboards/create": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Create dashboard",
                "parameters": [
                    {
                        "description": " ",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateOnePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dashboard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.CommonError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AddAccessPayload": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string"
                },
                "dashboard_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "httpext.CommonError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.Dashboard": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Field"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pipelines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Pipeline"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DashboardsUsers"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}